+++ sources/Vertex.cpp
@@ -108,7 +108,7 @@ Vertex Vertex::crossProduct(const Vertex &lhs, const Vertex &rhs) {
 
 void Vertex::parsestr(const std::string &pstr) {
 	std::string str = trim(pstr, " \t\n()");
-	size_t fspos = str.find_first_of(' '), espos = str.find_last_of(' ');
+	unsigned int fspos = str.find_first_of(' '), espos = str.find_last_of(' ');
 	if(fspos == std::string::npos || espos == std::string::npos || fspos == espos) {
 		vertex_[0] = std::numeric_limits<double>::quiet_NaN();
 		vertex_[1] = std::numeric_limits<double>::quiet_NaN();
